+++ tests/completion_e2e/test_completion_e2e.py
@@ -58,6 +58,8 @@ test_cases: Dict[str, str] = {
         "develop",
     "git machete github ":
         "anno-prs checkout-prs create-pr restack-pr retarget-pr sync",
+    "git machete github anno-prs -":
+        "--debug -h --help --with-urls -v --verbose",
     "git machete github checkout-prs -":
         "--all --by --debug -h --help --mine -v --verbose",
     "git machete github create-pr -":